moved the logic to a submodule
