+++ packages/simulator/src/profiles/standard/accounts.json
@@ -20,3 +20,4 @@
     "lastSyncDate": "1995-12-17T03:24:00"
   }
 ]
+
